test: split invite replay out of downtime-replay.test.ts
The other two tests in that file covered room_message replay via
applyStateSync/deliveryQueues -- machinery the directed fan-out no
longer uses for message delivery (see room-send-retry.integration.test.ts
for that behaviour's own replacement). Invite delivery itself is
untouched by that migration, so its own test keeps the legacy
replay path it actually exercises, just renamed to reflect that it's
now the only thing left in the file.
